Migrate billing repository to unit test
