+++ android/lib/billing/src/test/kotlin/android/content/Context.kt
@@ -1,0 +1,3 @@
+package android.content
+
+open class Context
